% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Checker.R
\docType{data}
\name{checker}
\alias{R6Checker}
\alias{checker}
\title{R6 Class of Standard Checking Utilities}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
checker()
}
\description{
Create a new R6 Class that contains public methods of standard checking utilities
}
\keyword{datasets}

